requirements.txt
Django==1.9.4
django-suit==0.2.18
django-watson==1.2.1
Pillow==3.1.1
psycopg2==2.6.1
South==1.0.2

#https://www.microsoft.com/en-us/download/details.aspx?id=44266